mypy

+++ backend/danswer/configs/model_configs.py
@@ -99,5 +99,5 @@ GEN_AI_MAX_OUTPUT_TOKENS = int(os.environ.get("GEN_AI_MAX_OUTPUT_TOKENS") or 102
 GEN_AI_MAX_INPUT_TOKENS = int(os.environ.get("GEN_AI_MAX_INPUT_TOKENS") or 3000)
 # Leave 20% other prompt pieces, this only applies to secondary LLM flows as the main chat
 # flow has separate logic to drop messages
-GEN_AI_HISTORY_CUTOFF = 0.8 * GEN_AI_MAX_INPUT_TOKENS
+GEN_AI_HISTORY_CUTOFF = int(0.8 * GEN_AI_MAX_INPUT_TOKENS)
 GEN_AI_TEMPERATURE = float(os.environ.get("GEN_AI_TEMPERATURE") or 0)
